payment/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
fastapi==0.104.0
h11==0.14.0
hiredis==2.2.3
httptools==0.6.1
idna==3.4
more-itertools==9.1.0
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
python-dotenv==1.0.0
python-ulid==1.1.0
PyYAML==6.0.1
redis==4.6.0
redis-om==0.2.1
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
types-pyOpenSSL==23.2.0.2
types-redis==4.6.0.7
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
watchfiles==0.21.0
websockets==12.0
